Code restructure and front end revamp
Code split into frontend and backend. Controllers and api are all in server folder. Fronted is in the client folder.

For frontend changes use reactjs and Typescript. Typescript is a superset of Javascript so basically different file extension, but you code in reactjs javascript. Typescript is object oriented so the rules are stricter to prevent errors before even running the code.